requirements.txt
pypdf
dotenv
transformers
numpy
sentence-transformers
google-genai
argparse